:white_check_mark: test(meta): 测试迁父级 tests/ 并清理 ignore 标记
- std 查找链与 yx 六模块测试迁入父级 tests/，由模块根声明（规则 1.1）
- 撤销 #[ignore] bless 入口，治愈改为 examples/gen-std-interfaces 工具
- 多断言测试表驱动化并补 Arrange/Act/Assert 三段（规则 4.1/4.2）
- 辅助函数 unwrap 补失败上下文（规则 5.2）
- 被测私有函数提权 pub(crate) 供父级测试注入
