+++ src/std/tests/gen_interfaces.rs
@@ -1,10 +1,18 @@
-//! 标准库接口文件生成器测试
+//! 标准库接口文件测试（RFC-037 §标准库目录）
 //!
-//! 验证 `gen_interfaces` 模块从 `StdModule` trait 自动生成 `.yx` 接口文件的功能。
+//! 覆盖 gen_interfaces.rs 的两块行为：
+//! - 接口视图生成（`StdModule::exports()` → `.yx` 签名视图，发行包
+//!   `lib/yaoxiang/std/` 的 native 层来源）
+//! - 三级查找链（项目 vendor → 发行包 exe 相对 → 全局回退）
+//! - 预生成文件同步门禁（生成物入库，漂移即红，同 RFC-013 码表模式）
+
+use tempfile::TempDir;
 
 use crate::std::gen_interfaces::{
-    generate_all_interfaces, write_interfaces_to_dir, find_std_interface_file,
+    find_std_interface_file, find_std_interface_file_in, generate_all_interfaces,
+    write_interfaces_to_dir,
 };
+use std::path::{Path, PathBuf};
 
 #[test]
 fn test_generate_all_interfaces() {
@@ -79,14 +87,113 @@ fn test_find_std_interface_file() {
     assert!(result.is_none());
 }
 
-/// 仓库预生成接口目录（RFC-037：发行包 lib/yaoxiang/std 的 native 层来源，
-/// 打包时纯复制，不再有运行时生成入口）
+/// 在 `dir` 下写一个接口文件，返回其路径（模拟查找链中的某一环）
+fn put_interface(
+    dir: &Path,
+    name: &str,
+) -> PathBuf {
+    std::fs::create_dir_all(dir).unwrap();
+    let file = dir.join(format!("{}.yx", name));
+    std::fs::write(&file, "// stub")
+        .unwrap_or_else(|e| panic!("写接口桩文件 {} 失败: {e}", file.display()));
+    file
+}
+
+#[test]
+fn test_find_project_override_beats_bundled_and_global() {
+    // Arrange: 三级链各放一个同名接口文件
+    let proj = TempDir::new().unwrap();
+    let pkg = TempDir::new().unwrap();
+    let global = TempDir::new().unwrap();
+    let project_file = put_interface(&proj.path().join(".yaoxiang/vendor/std"), "io");
+    put_interface(&pkg.path().join("lib/yaoxiang/std"), "io");
+    put_interface(global.path(), "io");
+
+    // Act
+    let found = find_std_interface_file_in(
+        Some(proj.path()),
+        Some(&pkg.path().join("bin")),
+        Some(global.path()),
+        "io",
+    );
+
+    // Assert: 项目覆盖必须优先于发行包与全局
+    assert_eq!(
+        found.as_deref(),
+        Some(project_file.as_path()),
+        "project vendor/std override must win over bundled and global"
+    );
+}
+
+#[test]
+fn test_find_exe_relative_bundled_std_resolves_pkg_layout() {
+    // Arrange: 只有发行包目录有该文件（bin/ 引擎 + lib/yaoxiang/std/）
+    let pkg = TempDir::new().unwrap();
+    let global = TempDir::new().unwrap();
+    let bundled = put_interface(&pkg.path().join("lib/yaoxiang/std"), "math");
+
+    // Act: exe 位于 <pkg>/bin/，接口应解析到 <pkg>/lib/yaoxiang/std/
+    let found = find_std_interface_file_in(
+        None,
+        Some(&pkg.path().join("bin")),
+        Some(global.path()),
+        "math",
+    );
+
+    // Assert
+    assert_eq!(
+        found.as_deref(),
+        Some(bundled.as_path()),
+        "exe-relative ../lib/yaoxiang/std must be found from bin/"
+    );
+}
+
+#[test]
+fn test_find_global_dir_is_last_fallback() {
+    // Arrange: 只有全局目录有该文件
+    let global = TempDir::new().unwrap();
+    let global_file = put_interface(global.path(), "time");
+
+    // Act
+    let found = find_std_interface_file_in(None, None, Some(global.path()), "time");
+
+    // Assert
+    assert_eq!(
+        found.as_deref(),
+        Some(global_file.as_path()),
+        "global dir must be used when project and bundled are absent"
+    );
+}
+
+#[test]
+fn test_find_unknown_module_returns_none() {
+    // Arrange: 空目录
+    let empty = TempDir::new().unwrap();
+
+    // Act
+    let found = find_std_interface_file_in(
+        Some(empty.path()),
+        Some(&empty.path().join("bin")),
+        Some(empty.path()),
+        "no_such_module",
+    );
+
+    // Assert
+    assert!(
+        found.is_none(),
+        "missing module must return None, got {:?}",
+        found
+    );
+}
+
+/// 仓库预生成接口目录（发行包 lib/yaoxiang/std 的 native 层来源，
+/// 打包时纯复制，无运行时生成入口）
 const COMMITTED_INTERFACES_DIR: &str = "src/std/interfaces";
 
 #[test]
 fn test_committed_interface_files_match_generation() {
     // Arrange: 预生成目录与 StdModule::exports() 是"单一源 → 派生文件"关系
-    // （同 RFC-013 码表：漂移即红，治愈走 bless 入口）
+    // （同 RFC-013 码表：漂移即红，治愈走 example 工具）
     let committed = std::path::Path::new(env!("CARGO_MANIFEST_DIR")).join(COMMITTED_INTERFACES_DIR);
     let expected = generate_all_interfaces();
 
@@ -95,14 +202,14 @@ fn test_committed_interface_files_match_generation() {
         let path = committed.join(format!("{}.yx", name));
         let on_disk = std::fs::read_to_string(&path).unwrap_or_else(|e| {
             panic!(
-                "预生成接口文件 {} 缺失: {}；治愈：cargo test update_committed_interface_files -- --ignored",
+                "预生成接口文件 {} 缺失: {}；治愈：cargo run --example gen-std-interfaces",
                 path.display(),
                 e
             )
         });
         assert_eq!(
             &on_disk, content,
-            "预生成接口 {} 与 StdModule::exports() 漂移；治愈：cargo test update_committed_interface_files -- --ignored",
+            "预生成接口 {} 与 StdModule::exports() 漂移；治愈：cargo run --example gen-std-interfaces",
             name
         );
     }
@@ -113,19 +220,8 @@ fn test_committed_interface_files_match_generation() {
         let file_name = entry.unwrap().file_name().to_string_lossy().to_string();
         assert!(
             valid.contains(&file_name),
-            "预生成目录存在孤儿文件 {}（生成器已不再产出）；治愈：删除后重跑 bless",
+            "预生成目录存在孤儿文件 {}（生成器已不再产出）；治愈：删除后重跑 gen-std-interfaces example",
             file_name
         );
     }
-}
-
-#[test]
-#[ignore = "bless 入口：std 接口变更后显式运行重写预生成文件（cargo test update_committed_interface_files -- --ignored）"]
-fn update_committed_interface_files() {
-    // Act: 以 StdModule::exports() 为唯一源重写仓库预生成目录
-    let dir = std::path::Path::new(env!("CARGO_MANIFEST_DIR")).join(COMMITTED_INTERFACES_DIR);
-    let count = write_interfaces_to_dir(&dir).unwrap();
-
-    // Assert: 数量可感知（bless 不做内容断言，正确性由同步测试把关）
-    assert!(count > 0, "bless 应至少写出一个接口文件");
 }
