+++ tools/yx/src/tests/mod.rs
@@ -1,0 +1,10 @@
+//! yx 前门测试（RFC-037）
+//!
+//! 按被测单文件模块分文件：dl / pin / platform / resolve / settings / toolchain。
+
+mod dl;
+mod pin;
+mod platform;
+mod resolve;
+mod settings;
+mod toolchain;
