+++ tools/yx/src/tests/pin.rs
